requirements.txt
google-genai
numpy
python-dotenv
